# docs/swagger.yaml
basePath: /
definitions:
...
  shop.Item:
    properties:
      brand:
        type: string
      category:
        description: '"штаны", "футболка", "кофта" и т.д.'
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      photo_url:
        description: ссылка на картинку
        type: string
      purchase_price:
        description: Цена закупки
        type: number
      sale_price:
        description: Цена продажи
        type: number
      shop_id:
        type: integer
      size:
        type: string
      updated_at:
        type: string
    type: object
  shop.Shop:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      owner_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API для CRM системы управления магазинами
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: CRM Backend API
  version: "1.0"
paths:
  /admin/shops:
    get:
      consumes:
      - application/json
      description: Get a list of all shops
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/shop.Shop'
            type: array
        "500":
          description: ошибка получения магазинов
          schema:
            type: string
      summary: Get all shops
      tags:
      - shops
    post:
      consumes:
      - application/json
      description: Create a new shop for a user
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Shop object
        in: body
        name: shop
        required: true
        schema:
          $ref: '#/definitions/shop.Shop'
      produces:
      - application/json
      responses:
        "201":
          description: Магазин создан для пользователя
          schema:
            type: string
        "400":
          description: неправильный формат данных
          schema:
            type: string
        "403":
          description: доступ запрещён
          schema:
            type: string
        "500":
          description: ошибка создания магазина
          schema:
            type: string
      summary: Create new shop
      tags:
      - shops
...
  /owner/shops:
    get:
      consumes:
      - application/json
      description: Get all shops belonging to the authenticated owner
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/shop.Shop'
            type: array
        "401":
          description: не авторизован
          schema:
            type: string
        "500":
          description: ошибка получения магазинов
          schema:
            type: string
      summary: Get owner's shops
      tags:
      - owner
...
  /owner/shops/{shopID}/items:
    get:
      consumes:
      - application/json
      description: Get all items for a specific shop
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Shop ID
        in: path
        name: shopID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/shop.Item'
            type: array
        "400":
          description: invalid shopID
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: Get all items in a shop
      tags:
      - items
    post:
      consumes:
      - application/json
      description: Create a new item in a shop
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Shop ID
        in: path
        name: shopID
        required: true
        type: integer
      - description: Item details
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/shop.Item'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/shop.Item'
        "400":
          description: invalid request body
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: Create new item
      tags:
      - items
  /owner/shops/{shopID}/items/{itemID}:
    delete:
      consumes:
      - application/json
      description: Delete an item by ID
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Shop ID
        in: path
        name: shopID
        required: true
        type: integer
      - description: Item ID
        in: path
        name: itemID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid itemID
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: Delete item
      tags:
      - items
    get:
      consumes:
      - application/json
      description: Get a specific item by ID
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Shop ID
        in: path
        name: shopID
        required: true
        type: integer
      - description: Item ID
        in: path
        name: itemID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shop.Item'
        "400":
          description: invalid itemID
          schema:
            type: string
        "404":
          description: error message
          schema:
            type: string
      summary: Get specific item
      tags:
      - items
    put:
      consumes:
      - application/json
      description: Update an existing item
      parameters:
      - description: Bearer JWT token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Shop ID
        in: path
        name: shopID
        required: true
        type: integer
      - description: Item ID
        in: path
        name: itemID
        required: true
        type: integer
      - description: Item details
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/shop.Item'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid request body
          schema:
            type: string
        "500":
          description: error message
          schema:
            type: string
      summary: Update item
      tags:
      - items
schemes:
- http
swagger: "2.0"